refactor: 继续模块化灵动岛生命周期接口
