+++ src/main/ipc/app.ts
@@ -1,6 +1,48 @@
 import { app, ipcMain, shell } from 'electron';
+import { basename } from 'path';
+import { clearLogsCacheFiles, ensureLogsDir } from '../log/mainLog';
 
 export function registerAppIpcHandlers(): void {
+  ipcMain.on('app:quit', () => {
+    app.quit();
+  });
+
+  ipcMain.handle('app:restart', () => {
+    try {
+      app.relaunch();
+      app.exit(0);
+      return true;
+    } catch (err) {
+      console.error('[App] restart error:', err);
+      return false;
+    }
+  });
+
+  ipcMain.handle('app:open-logs-folder', async () => {
+    try {
+      const logDir = ensureLogsDir();
+      const result = await shell.openPath(logDir);
+      return result === '';
+    } catch (err) {
+      console.error('[App] open logs folder error:', err);
+      return false;
+    }
+  });
+
+  ipcMain.handle('app:clear-logs-cache', async () => {
+    try {
+      const result = clearLogsCacheFiles();
+      if (!result.success) {
+        return { success: false, freedBytes: 0 };
+      }
+      console.log(`[App] cleared logs cache: ${result.fileCount} files, ${(result.freedBytes / 1024).toFixed(1)} KB freed`);
+      return { success: true, freedBytes: result.freedBytes };
+    } catch (err) {
+      console.error('[App] clear logs cache error:', err);
+      return { success: false, freedBytes: 0 };
+    }
+  });
+
   ipcMain.handle('app:get-file-icon', async (_event, filePath: string) => {
     try {
       let iconPath = filePath;
@@ -27,6 +69,19 @@ export function registerAppIpcHandlers(): void {
     } catch (err) {
       console.error('[App] open-file error:', err);
       return false;
+    }
+  });
+
+  ipcMain.handle('app:resolve-shortcut', (_event, lnkPath: string) => {
+    try {
+      if (process.platform === 'win32') {
+        const result = shell.readShortcutLink(lnkPath);
+        return { target: result.target, name: basename(lnkPath, '.lnk') };
+      }
+      return null;
+    } catch (err) {
+      console.error('[App] resolve-shortcut error:', err);
+      return null;
     }
   });
 }
